Perform tests to verify the functionality
